+++ core/Player.py
@@ -1,4 +1,4 @@
-class Jugador:
+class Player:
     def __init__(self, numero, ficha):
         self.numero = numero  # Jugador 1 o Jugador 2
         self.ficha = ficha    # 'X' o 'O'